+++ pykechain/models/notification.py
@@ -6,7 +6,7 @@ from pykechain.models import Base
 from typing import Text, List, Optional, Dict, Union
 
 from pykechain.models.input_checks import check_text, check_enum, check_base, check_user
-from pykechain.utils import is_valid_email
+from pykechain.utils import is_valid_email, Empty, clean_empty_values, empty
 
 
 class Notification(Base):
@@ -76,38 +76,47 @@ class Notification(Base):
 
     def edit(
             self,
-            subject: Optional[Text] = None,
-            message: Optional[Text] = None,
-            status: Optional[NotificationStatus] = None,
-            recipients: Optional[List[Union['User', Text, int]]] = None,
-            team: Optional[Union['Team', Text]] = None,
-            from_user: Optional[Union['User', Text]] = None,
-            event: Optional[NotificationEvent] = None,
-            channel: Optional[NotificationChannels] = None,
+            subject: Optional[Union[Text, Empty]] = empty,
+            message: Optional[Union[Text, Empty]] = empty,
+            status: Optional[Union[NotificationStatus, Empty]] = empty,
+            recipients: Optional[Union[List[Union['User', Text, int]], Empty]] = empty,
+            team: Optional[Union['Team', Text, Empty]] = empty,
+            from_user: Optional[Union['User', Text, Empty]] = empty,
+            event: Optional[Union[NotificationEvent, Empty]] = empty,
+            channel: Optional[Union[NotificationChannels, Empty]] = empty,
             **kwargs
     ) -> None:
         """
         Update the current `Notification` attributes.
 
-        :param subject: (O) Header text of the notification
-        :type subject: str
-        :param message: (O) Content message of the notification
-        :type message: str
-        :param status: (O) life-cycle status of the notification, defaults to "DRAFT".
+        Setting an input to None will clear out the value (only applicable to recipients and from_user).
+
+        :param subject: (O) Header text of the notification. Cannot be cleared.
+        :type subject: basestring or None or Empty
+        :param message: (O) Content message of the notification. Cannot be cleared.
+        :type message: basestring or None or Empty
+        :param status: (O) life-cycle status of the notification, defaults to "DRAFT". Cannot be cleared.
         :type status: NotificationStatus
         :param recipients: (O) list of recipients, each being a User object, user ID or an email address.
-        :type recipients: list
+        :type recipients: list or None or Empty
         :param team: (O) team object to which the notification is constrained
         :type team: Team object or Team UUID
         :param from_user: (O) Sender of the notification, either a User object or user ID. Defaults to script user.
-        :type from_user: User or user ID
-        :param event: (O) originating event of the notification.
+        :type from_user: User or user ID or None or Empty
+        :param event: (O) originating event of the notification. Cannot be cleared.
         :type event: NotificationEvent
-        :param channel: (O) method used to send the notification, defaults to "EMAIL".
+        :param channel: (O) method used to send the notification, defaults to "EMAIL". Cannot be cleared.
         :type channel: NotificationChannels
         :param kwargs: (optional) keyword=value arguments
         :return: None
         :raises: APIError: when the `Notification` could not be updated
+
+        Not mentioning an input parameter in the function will leave it unchanged. Setting a parameter as None will
+        clear its value (where that is possible). The example below will clear the from_user, but leave everything
+        else unchanged.
+
+        >>> notification.edit(from_user=None)
+
         """
         from pykechain.models import Team, User
 
@@ -121,24 +130,36 @@ class Notification(Base):
                         recipient_emails.append(recipient)
                     else:
                         recipient_users.append(check_user(recipient, User, 'recipient'))
+            elif isinstance(recipients, Empty):
+                recipient_emails = empty
+                recipient_users = empty
             else:
                 raise IllegalArgumentError('`recipients` must be a list of User objects, IDs or email addresses, '
                                            '"{}" ({}) is not.'.format(recipients, type(recipients)))
 
+        if isinstance(channel, Empty):
+            channels = empty
+        elif check_enum(channel, NotificationChannels, 'channel'):
+            channels = [channel]
+        else:
+            channels = list()
+
         update_dict = {
-            'status': check_enum(status, NotificationStatus, 'status'),
-            'event': check_enum(event, NotificationEvent, 'event'),
-            'subject': check_text(subject, 'subject'),
-            'message': check_text(message, 'message'),
+            'status': check_enum(status, NotificationStatus, 'status') or self.status,
+            'event': check_enum(event, NotificationEvent, 'event') or self.event,
+            'subject': check_text(subject, 'subject') or self.subject,
+            'message': check_text(message, 'message') or self.message,
             'recipient_users': recipient_users,
             'recipient_emails': recipient_emails,
             'team': check_base(team, Team, 'team'),
             'from_user': check_user(from_user, User, 'from_user'),
-            'channels': [channel] if check_enum(channel, NotificationChannels, 'channel') else [],
+            'channels': channels or self.channels
         }
 
         if kwargs:
             update_dict.update(kwargs)
+
+        update_dict = clean_empty_values(update_dict=update_dict)
 
         url = self._client._build_url('notification', notification_id=self.id)